Fix cap'ing PCRs and ensure loclaity 2 avail on shutdown
Since the kernel may have used the TPM and not released the locality it used, we need to ensure that localities 0 and 1 are released before cap'ing the PCRs using locality 2, else that will fail.

We need to cap all dynamic PCRs that were extended as part of the launch (17, 18, and any in the policy).  Static PCRs in the policy do not get cap'ed, since during S3 they do not get reset like the dynamic PCRs do.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>