Remove the shutdown_entry32 and shutdown_entry64 from tboot_shared_t and have just a single shutdown_entry field.
This requires a change to the tboot_shared_t data structure that is not backwards compatible, so the version has been incremented to 3.
Since backwards compatbility was broken with the previous patch, this is intended to be used with a set of patches to Xen that support the new tboot_shared_t.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>