Create common fn to determine end addr of tboot; don't set boot_params->tboot_shared_addr if launching Linux w/o TXT
Create a common fn to determine the end addr of tboot, as marked E820_RESERVED in the e820 table and used to locate the beginning of a relocatable kernel.

If there was an error during launch that prevents the measured environment from launching, but allows Linux to boot, don't set the boot_params->tboot_shared_addr field so that Linux will know that there was no measured launch.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>