Fix build errors with gcc v4.1.2
The -Wlogical-op option is not supported in gcc v4.1.2 so make its use condition on support.

The -Wextra flag caused a warning in paging.c, which was disabled by #pragma GCC diagnostic ignored "-Wtype-limits".  However, that pragma is not supported in gcc v4.1.2 and so in order to prevent the warning, the comparison has been removed and a compile-time assertion added to ensure this is safe.

Signed-off-by: Joseph Cihula <joseph.cihula@intel.com>
ACKed-by: Xudong Hao <xudong.hao@intel.com>