Make lcp_crtpollist support empty lists (i.e. with no elements)
Per the MLE Developers Guide, an empty policy list is allowed as a placeholder to be updated before the first use.  This allows for determining the policy hash (assuming the list is signed) ahead of time so that it can be provisioned in the policy in TPM NV.

This patch fixes lcp_crtpollist to allow this.

Signed-off-by:  Joseph Cihula <joseph.cihula@intel.com>