Fix compile issue when debug=y
I found a problem when compiling tboot 1.8.1 on SUSE 13.1 with gcc 4.8.1 - when I switched on debug i.e.
export debug=y
I got some bizarre messages about trousers not being installed - it was indeed installed and it took me a while to figure out that the real problem was the definition of the CFLAGS_WARN for tboot below is a patch that fixes it for me.
The bug is still there in 1.8.2
....JW

Signed-off-by: John Marland <johnwmarland@gmail.com>
Signed-off-by: Gang Wei <gang.wei@intel.com>