Access tpm sts reg with 3-byte width in v1.2 case and 4-byte width in v2.0 case
Tboot booting will freeze when using tboot-1.8.1 as provided by the CentOS
distribution on below system.
1. Dell Latitude e6540 laptop
2. Dell Venue 7140

Via code review, there found an issue with how tboot handles writing/reading
the status registers while trying to handle both the TPM 1.2 and 2.0 cases.
The definition of the status register is always 4 bytes long(as defined in
"tboot/common/tpm.c", however TPM 1.2 devices still only have a 3 byte register.
This causes writes/reads to access memory that they should not be accessing
when communicating with 1.2 versioned TPMs.

This patch create 2 different structures for the status register and use one or
the other based on which type of TPM is present.

This patch solved the problem on the listed two systems.

Signed-off-by: Brent Collins <bcollins@trustedcs.com>
Signed-off-by: Gang Wei <gang.wei@intel.com>