We found this problem when booting a KVM guest through tboot from a host OS where the VMX register is not exposed to the guest, even when the guest has cloned the host CPU.
Attempting to read MSR_IA32_FEATURE_CONTROL before checking whether
it exists, on CPUs where it does not exist, sends the BSP into an
infinite loop. #GP is asserted when trying to read the non-existent
MSR, which resets the IP, only to again encounter the attempted read
of the non-existent MSR.

Postponing the read of MSR_IA32_FEATURE_CONTROL until the existence
of VMX/SMX has been ascertained prevents this problem.

Signed-off-by: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Bandan Das <bsd@redhat.com>
Signed-off-by: Ning Sun <ning.sun@intel.com>