Fix the evtlog dump function for tpm2 case
The TXT MLE DG appendix G specified that for non-sha1 log, the first
event shall be a tpm1.2 format event. The original code didn't treat
this requirement accordingly, so that it will fail to dump any non-sha1
logs.

This also fixed the issue that the hash values in tboot created event
log were always mis-written with 4 bytes of 0s.

Signed-off-by: Gang Wei <gang.wei@intel.com>
Signed-off-by: Ning Sun <ning.sun@intel.com>