Check for initrd image before moving.
Currently, TBoot attempts to move the initrd image without checking
if an initrd image is present. If no initrd image is present, Linux
expects the base address of the initrd image to be set to zero in
the Linux header.

The patch modifies the Linux loading code such that the initrd image
is only moved if the size is greater than zero.

Signed-off-by: Safayet Ahmed <Safayet.Ahmed@ge.com>
Reviewed-by: Ning Sun <ning.sun@intel.com>