Build binary also in PE format
GRUB2 allows to check signature of application loaded by multiboot2 using
verifies framework. That functionality requires PE format, signed in the same
way as for SecureBoot.

This patch introduces tboot.mb2 file that is a PE binary that can be signed and
loaded via multiboot2. Keep in mind that although PE header indicates that this
is a 64bit application, under the hood there is a standard 32bit TBOOT version that
can't be run without GRUB2.

GRUB2 2.06 is required to use tboot.mb2

Signed-off-by: Lukasz Hawrylko <lukasz.hawrylko@intel.com>