Drop support for Linux kernel in ELF format
Every modern distro has Linux kernel as bzImage, so ELF support is no longer
needed. After that change there is no need to replace E820 in mbi structure
with modified one that marks few memory regions as reserved. Problem was that
modified E820 is bigger than original one and we can't be sure that there is
enough space for it in mbi.

When using bzImage, E820 is passed to kernel by boot params, not by mbi, so
there is more flexibility to put it somewhere in available memory.

Signed-off-by: Lukasz Hawrylko <lukasz.hawrylko@intel.com>